Add support for multi-value proposal
Use the `--enable-multi-value` flag to enable.

A lot of code already "worked" with multi-value, and just needed to
remove the restrictions. Most of the other changes are modifying the
callback APIs to be more general, e.g. taking more than 1 result type.

* Types are now stored as the negative values; this works nicely with
  the encoding of inline function types (used for block signatures),
  which are always positive values.
* Remove `BlockSignature` and use `BlockDeclaration` instead, which
  is just a typedef to `FuncSignature`. This allows for explicit or
  implicit type specifications on the block signatures.
* Allow for >1 "keep" values in the DropKeep interpreter instruction

+++ src/interp.h
@@ -85,10 +85,10 @@ static const IstreamOffset kInvalidIstreamOffset = ~0;
 //   struct {
 //     IstreamOffset offset;
 //     uint32_t drop_count;
-//     uint8_t keep_count;
+//     uint32_t keep_count;
 //   };
 #define WABT_TABLE_ENTRY_SIZE \
-  (sizeof(IstreamOffset) + sizeof(uint32_t) + sizeof(uint8_t))
+  (sizeof(IstreamOffset) + sizeof(uint32_t) + sizeof(uint32_t))
 #define WABT_TABLE_ENTRY_OFFSET_OFFSET 0
 #define WABT_TABLE_ENTRY_DROP_OFFSET sizeof(IstreamOffset)
 #define WABT_TABLE_ENTRY_KEEP_OFFSET (sizeof(IstreamOffset) + sizeof(uint32_t))
@@ -522,7 +522,7 @@ class Thread {
   template <typename T>
   ValueTypeRep<T> PopRep();
 
-  void DropKeep(uint32_t drop_count, uint8_t keep_count);
+  void DropKeep(uint32_t drop_count, uint32_t keep_count);
 
   Result PushCall(const uint8_t* pc) WABT_WARN_UNUSED;
   IstreamOffset PopCall();
